Import component updates - WIP
Committing imput component updates for review. These are very
work-in-progress, but I would like to finalise this week.
