+++ Structure_Alligator/Structural/Properties/Constraints/NodeConstraint.cs
@@ -21,7 +21,7 @@ namespace BH.UI.Alligator.Structural.Properties
                 return new Guid("5FE0E2C4-5E10-410F-BBC7-C255FE1BD2B3");
             }
         }
-
+        
         /// <summary> Icon (24x24 pixels)</summary>
         protected override System.Drawing.Bitmap Internal_Icon_24x24
         {
